refactor(deletion_policy): move to separate file and add to service config
